disable firewalla upgrade if firerouter no auto upgrade is set

+++ scripts/fireupgrade.sh
@@ -193,6 +193,12 @@ if [[ -e "/home/pi/.firewalla/config/.no_auto_upgrade" ]]; then
   exit 0
 fi
 
+if [[ -e "/home/pi/.router/config/.no_auto_upgrade" ]]; then
+  $FIRELOG -t debug -m "FIREWALLA.UPGRADE NO UPGRADE -- ON FIREROUTER"
+  echo '======= SKIP UPGRADING BECAUSE OF FIREROUTER FLAG /home/pi/.router/config/.no_auto_upgrade ======='
+  exit 0
+fi
+
 if $(/bin/systemctl -q is-active watchdog.service) ; then sudo /bin/systemctl stop watchdog.service ; fi
 sudo rm -f $FIREWALLA_HOME/.git/*.lock
 # ensure the remote fetch branch is up-to-date
